Enable datetime-local
I’ve not updated Chrome’s status. Need to work out what is going on
there, or updated score (need to discuss with you @stevefaulkner to see
how that is worked out)

Left Edge as unknown as all the rest are. In reality it is supported in
Edge 13. Works with keyboard. Has some hooking up to a11y API but we’re
revisiting in upcoming version to see what makes sense. The spec says
calendar role, which doesn’t make sense for our implementation (we have
spinners, not a non-touch friendly small calendar grid)