Preliminary markup for results section
Kept the logos in for now so they don’t go away, but should probably be
CSS? If keep in markup, could use flex box to reorder, or a float.

I put browser name first as has to be read out by screen reader first.
Can use flex box or other tricks to re-order, although once browsers
read in layout order, that would break.

Not exactly sure if names should be H3s, and others as p’s or not.
Could be spans, etc. Reason I didn’t use H2 is the non-existing header
of the section should be h2 (using an aria-label instead)