Updated most of the margin/paddings to redlines
Except header as it looks bad, and icons as they’re bg images and need
the final ones in first. Table cells are especially wider as the icons
are centred so there is no margin/padding to apply the 12px to. need to
get the icons at the correct size and try to bypass how it is ignoring
the max-width, but not much point doing that yet.