Make table cells closer to correct height
Tables can expand depending on content so not exactly right (and a few
pixels smaller as it ended up looking too tall for some reason.

Can’t use height as that increases the padding and thus breaks the
dotted border. Not sure why when the padding is explicitly called out
as 8px, but must be a table thing.