Make dotted border work via border-image
@melanierichards …I never want to have to do that ever again, in a
month of sundays. OMG.

All the margins and sizes are probably wrong in mobile and tablet now
as I had to change how everything was done and I don’t remember all the
expanded sizes, etc. Vertical align that was working might not be
exactly correct now either. Need to take a look again when my brain
clears.