Hide rows instead of cells, and edit colspans to match
Col spans with hidden rows or cells was making VO bail out when reading
full page. Seems like a bug as it should be intelligent enough to know
if display: none or aria-hidden is applied to rows that the colspan
applies to.

Either way, I’ve added more complex logic to add hidden attributes to
rows, and remove the colspan, saving in a data-colspan attribute. and
adding/removing when expand/collapse.

Also tidied up CSS to not use expand/collapse classes and not hide rows
if JS hasn’t run.

Fixes #21